refactor before merge
