+++ app/(dashboard)/page.tsx
@@ -1,10 +1,8 @@
 'use client'
 import React from 'react';
-import SubBarNavTop from './subBarNav/SubBarNavTop';
 const Dashboard = () => {
   return <div>
     Dashboard
-    <SubBarNavTop/>
   </div>;
 };
